requirements.txt
fastapi~=0.115.11
uvicorn~=0.34.0
mysql-connector-python~=9.2.0
dotenv~=0.9.9
sqlmodel~=0.0.23
pydantic~=2.10.6
pydantic-settings~=2.8.1
mysql-connector-python==9.2.0
boto3~=1.37.7
pytest~=8.2.2
h11~=0.14.0
pip~=24.0
filelock~=3.17.0
typing_extensions~=4.12.2
sniffio~=1.3.1
setuptools~=65.5.0
click~=8.1.7
starlette~=0.46.0
pydantic_core~=2.27.2
anyio~=4.8.0
annotated-types~=0.7.0
SQLAlchemy~=2.0.38
PyYAML~=6.0.1
Jinja2~=3.1.4
packaging~=24.1
platformdirs~=4.3.6
Pygments~=2.18.0
python-dotenv~=1.0.1
python-multipart==0.0.20